COSTOR-1050: S3 Calling stop timer of metric multipart_create_object_success              at right place
Addition of comments to s3stats-whitelist.yaml and stopping
timer for metric multipart_create_object_success at right place

+++ server/s3_stats.cc
@@ -53,7 +53,7 @@ int S3Stats::count(const std::string& key, int value, int retry,
   return form_and_send_msg(key, "c", std::to_string(value), retry, sample_rate);
 }
 
-int S3Stats::timing(const std::string& key, int time_ms, int retry,
+int S3Stats::timing(const std::string& key, size_t time_ms, int retry,
                     float sample_rate) {
   return form_and_send_msg(key, "ms", std::to_string(time_ms), retry,
                            sample_rate);
